% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoClust}
\alias{CoClust}
\title{C++ code to to partition a bipartite network into non-overlapping biclusters, by optimizing bipartite modularity.}
\usage{
CoClust(nr, nc, data, ITER)
}
\arguments{
\item{nr}{Number of rows of incidence matrix.}

\item{nc}{Number of columns of incidence matrix.}

\item{data}{Vectorized incidence matrix.}

\item{ITER}{Number of iterations.}
}
\value{
MODULARITY Modularity value.

ASSIGN Partition of rows and columns.
}
\description{
C++ code to to partition a bipartite network into non-overlapping biclusters, by optimizing bipartite modularity.
}
