% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset of a bipartite network.}
\format{
A data frame with 798 rows and 8 binary variables:
\describe{
  \item{Symptom_1}{}
  \item{Symptom_2}{}
  \item{Symptom_3}{}
  \item{Symptom_4}{}
  \item{Symptom_5}{}
  \item{Symptom_6}{}
  \item{Symptom_7}{}
  \item{Symptom_8}{}
}
}
\usage{
example_data
}
\description{
A dataset consisting of anonymized patients (n=798) and symptoms (d=8), where each patient has one or more symptoms.
}
\keyword{datasets}
