% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset of a bipartite network.}
\format{A data frame with 798 rows and 8 binary variables:
\describe{
  \item{not.happy}{}
  \item{didnot.enjoyed.life}{}
  \item{felt.depressed}{}
  \item{everything.effort}{}
  \item{felt.lonely}{}
  \item{felt.sad}{}
  \item{couldnt.get.going}{}
  \item{sleep.restless}{}
}}
\usage{
example_data
}
\description{
A dataset consisting of stroke patients (n=798) and symptoms (d=8), where each patient has one or more symptoms.
}
\keyword{datasets}
